+++ alpha/apps/kaltura/lib/batch2/model/kConcatJobData.php
@@ -138,7 +138,7 @@ class kConcatJobData extends kJobData
 	}
 
 	/**
-	 * @return bool $sortNeeded
+	 * @return bool $shouldSort
 	 */
 	public function getShouldSort()
 	{