+++ alpha/apps/kaltura/lib/reports/kKavaReportsMgr.class.php
@@ -25,6 +25,7 @@ class kKavaReportsMgr extends kKavaBase
 	const METRIC_COUNT_TOTAL = 'count_total';
 	const METRIC_COUNT_TOTAL_ALL_TIME = 'count_total_all_time';
 	const METRIC_BANDWIDTH_SIZE_MB = 'bandwidth_consumption';
+	const METRIC_BANDWIDTH_SIZE_KB = 'bandwidth_consumption_kb';
 	const METRIC_TRANSCODING_SIZE_MB = 'transcoding_consumption';
 	const METRIC_STORAGE_ADDED_MB = 'added_storage';
 	const METRIC_STORAGE_DELETED_MB = 'deleted_storage';
@@ -51,6 +52,7 @@ class kKavaReportsMgr extends kKavaBase
 	const METRIC_BANDWIDTH_STORAGE_MB = 'combined_bandwidth_storage';
 	const METRIC_AVERAGE_STORAGE_AGGR_MONTHLY_MB = 'aggregated_monthly_avg_storage';
 	const METRIC_BANDWIDTH_STORAGE_AGGR_MONTHLY_MB = 'combined_bandwidth_aggregated_storage';
+	const METRIC_PEAK_ENTRIES = 'peak_entries';
 	
 	/// report settings
 	// report settings - common
@@ -78,10 +80,7 @@ class kKavaReportsMgr extends kKavaBase
 	const REPORT_FORCE_TOTAL_COUNT = 'report_force_total_count';
 	const REPORT_TABLE_MAP = 'report_table_map';
 	const REPORT_TABLE_FINALIZE_FUNC = 'report_table_finalize_func';
-	
-	const REPORT_LEGACY = 'report_legacy';
-	const REPORT_LEGACY_MAPPING = 'legacy_mapping';
-	const REPORT_LEGACY_JOIN_FIELD = 'join_field';
+	const REPORT_EDIT_FILTER_FUNC = 'report_edit_filter_func';
 	
 	// report settings - graph
 	const REPORT_GRANULARITY = 'report_granularity';
@@ -134,6 +133,30 @@ class kKavaReportsMgr extends kKavaBase
 	
 	const COLUMN_FORMAT_QUOTE = 'quote';
 	const COLUMN_FORMAT_UNIXTIME = 'unixtime';
+	
+	protected static $kava_enabled_reports = array(
+		myReportsMgr::REPORT_TYPE_TOP_CONTENT,
+		myReportsMgr::REPORT_TYPE_CONTENT_DROPOFF,
+		myReportsMgr::REPORT_TYPE_CONTENT_INTERACTIONS,
+		myReportsMgr::REPORT_TYPE_MAP_OVERLAY,
+		myReportsMgr::REPORT_TYPE_TOP_SYNDICATION,
+		myReportsMgr::REPORT_TYPE_USER_ENGAGEMENT,
+		myReportsMgr::REPORT_TYPE_SPECIFIC_USER_ENGAGEMENT,
+		myReportsMgr::REPORT_TYPE_USER_TOP_CONTENT,
+		myReportsMgr::REPORT_TYPE_USER_CONTENT_DROPOFF,
+		myReportsMgr::REPORT_TYPE_USER_CONTENT_INTERACTIONS,
+		myReportsMgr::REPORT_TYPE_APPLICATIONS,
+		myReportsMgr::REPORT_TYPE_PLATFORMS,
+		myReportsMgr::REPORT_TYPE_OPERATING_SYSTEM,
+		myReportsMgr::REPORT_TYPE_BROWSERS,
+		myReportsMgr::REPORT_TYPE_LIVE,
+		myReportsMgr::REPORT_TYPE_TOP_PLAYBACK_CONTEXT,
+		myReportsMgr::REPORT_TYPE_VPAAS_USAGE,
+	);
+		
+	protected static $kava_forced_reports = array(
+		myReportsMgr::REPORT_TYPE_ENTRY_USAGE,
+	);
 	
 	protected static $reports_def = array(
 		myReportsMgr::REPORT_TYPE_TOP_CONTENT => array(
@@ -299,12 +322,83 @@ class kKavaReportsMgr extends kKavaBase
 		),
 
 		myReportsMgr::REPORT_TYPE_VPAAS_USAGE => array(
-			self::REPORT_GRANULARITY => self::DRUID_GRANULARITY_MONTH,
-			self::REPORT_DIMENSION => self::DIMENSION_PARTNER_ID,
-			self::REPORT_METRICS => array(self::EVENT_TYPE_PLAY),
-			self::REPORT_LEGACY => array(
-				self::REPORT_LEGACY_MAPPING => array('total_plays' => 'count_plays'),
-				self::REPORT_LEGACY_JOIN_FIELD => 'month_id'),
+			self::REPORT_JOIN_REPORTS => array(
+				// bandwidth
+				array(
+					self::REPORT_DATA_SOURCE => self::DATASOURCE_BANDWIDTH_USAGE,
+					self::REPORT_FILTER_DIMENSION => self::DIMENSION_PARTNER_ID,
+					self::REPORT_GRAPH_METRICS => array(self::METRIC_BANDWIDTH_SIZE_MB),
+				),
+
+				// transcoding
+				array(
+					self::REPORT_DATA_SOURCE => self::DATASOURCE_TRANSCODING_USAGE,
+					self::REPORT_FILTER_DIMENSION => self::DIMENSION_PARTNER_ID,
+					self::REPORT_GRAPH_METRICS => array(self::METRIC_TRANSCODING_SIZE_MB),
+				),
+
+				// storage
+				array(
+					self::REPORT_DATA_SOURCE => self::DATASOURCE_STORAGE_USAGE,
+					self::REPORT_FILTER => array(		// can exclude logical deltas in this report
+						self::DRUID_DIMENSION => self::DIMENSION_EVENT_TYPE,
+						self::DRUID_VALUES => array(self::EVENT_TYPE_STATUS, self::EVENT_TYPE_PHYSICAL_ADD, self::EVENT_TYPE_PHYSICAL_DELETE)
+					),
+					self::REPORT_FILTER_DIMENSION => self::DIMENSION_PARTNER_ID,
+					self::REPORT_GRAPH_METRICS => array(self::METRIC_STORAGE_ADDED_MB, self::METRIC_STORAGE_DELETED_MB),
+				),
+
+				array(
+					self::REPORT_DATA_SOURCE => self::DATASOURCE_STORAGE_USAGE,
+					self::REPORT_INTERVAL => self::INTERVAL_BASE_TO_START,
+					self::REPORT_FILTER => array(		// can exclude logical deltas in this report
+						self::DRUID_DIMENSION => self::DIMENSION_EVENT_TYPE,
+						self::DRUID_VALUES => array(self::EVENT_TYPE_STATUS, self::EVENT_TYPE_PHYSICAL_ADD, self::EVENT_TYPE_PHYSICAL_DELETE)
+					),
+					self::REPORT_FILTER_DIMENSION => self::DIMENSION_PARTNER_ID,
+					self::REPORT_GRAPH_METRICS => array(self::METRIC_STORAGE_TOTAL_MB),
+					self::REPORT_GRAPH_ACCUMULATE_FUNC => 'self::addAggregatedStorageGraphs',
+				),
+
+				// media entries
+				array(
+					self::REPORT_DATA_SOURCE => self::DATASOURCE_ENTRY_LIFECYCLE,
+					self::REPORT_FILTER => array(
+						self::DRUID_DIMENSION => self::DIMENSION_MEDIA_TYPE,
+						self::DRUID_VALUES => array(self::MEDIA_TYPE_VIDEO, self::MEDIA_TYPE_AUDIO, self::MEDIA_TYPE_IMAGE, self::MEDIA_TYPE_FLASH)
+					),
+					self::REPORT_FILTER_DIMENSION => self::DIMENSION_PARTNER_ID,
+					self::REPORT_GRAPH_METRICS => array(self::METRIC_ENTRIES_ADDED, self::METRIC_ENTRIES_DELETED),
+				),
+
+				array(
+					self::REPORT_DATA_SOURCE => self::DATASOURCE_ENTRY_LIFECYCLE,
+					self::REPORT_INTERVAL => self::INTERVAL_BASE_TO_START,
+					self::REPORT_FILTER => array(
+						self::DRUID_DIMENSION => self::DIMENSION_MEDIA_TYPE,
+						self::DRUID_VALUES => array(self::MEDIA_TYPE_VIDEO, self::MEDIA_TYPE_AUDIO, self::MEDIA_TYPE_IMAGE, self::MEDIA_TYPE_FLASH)
+					),
+					self::REPORT_FILTER_DIMENSION => self::DIMENSION_PARTNER_ID,
+					self::REPORT_GRAPH_METRICS => array(self::METRIC_ENTRIES_TOTAL),
+					self::REPORT_GRAPH_ACCUMULATE_FUNC => 'self::addAggregatedEntriesGraphs',
+					),
+
+				// plays
+				array(
+					self::REPORT_FILTER_DIMENSION => self::DIMENSION_PARTNER_ID,
+					self::REPORT_GRAPH_METRICS => array(self::EVENT_TYPE_PLAY),
+				),
+					
+				// TODO: add total users
+			),
+			self::REPORT_GRAPH_AGGR_FUNC => 'self::aggregateUsageData',
+			self::REPORT_TABLE_MAP => array(
+				'total_plays' => 'count_plays',
+				'bandwidth_consumption' => self::METRIC_BANDWIDTH_SIZE_MB,
+				'average_storage' => self::METRIC_AVERAGE_STORAGE_MB,
+				'transcoding_consumption' => self::METRIC_TRANSCODING_SIZE_MB,
+				'total_media_entries' => self::METRIC_PEAK_ENTRIES,
+			),
 		),
 
 		myReportsMgr::REPORT_TYPE_TOP_CONTRIBUTORS => array(
@@ -544,7 +638,33 @@ class kKavaReportsMgr extends kKavaBase
 			self::REPORT_GRAPH_AGGR_FUNC => 'self::aggregateUsageData',
 			self::REPORT_GRAPH_FINALIZE_FUNC => 'self::addCombinedUsageGraph',
 		),
-			
+
+		myReportsMgr::REPORT_TYPE_ENTRY_USAGE => array(
+			self::REPORT_JOIN_REPORTS => array(
+				array(
+					self::REPORT_DATA_SOURCE => self::DATASOURCE_ENTRY_LIFECYCLE,
+					self::REPORT_FILTER => array(
+						self::DRUID_DIMENSION => self::DIMENSION_EVENT_TYPE,
+						self::DRUID_VALUES => array(self::EVENT_TYPE_STATUS, self::EVENT_TYPE_PHYSICAL_ADD, self::EVENT_TYPE_PHYSICAL_DELETE)
+					),
+					self::REPORT_FILTER_DIMENSION => self::DIMENSION_PARTNER_ID,
+					self::REPORT_GRAPH_METRICS => array(self::METRIC_ENTRIES_ADDED, self::METRIC_ENTRIES_DELETED),
+				),
+				array(
+					self::REPORT_DATA_SOURCE => self::DATASOURCE_ENTRY_LIFECYCLE,
+					self::REPORT_INTERVAL => self::INTERVAL_BASE_TO_START,
+					self::REPORT_FILTER => array(
+						self::DRUID_DIMENSION => self::DIMENSION_EVENT_TYPE,
+						self::DRUID_VALUES => array(self::EVENT_TYPE_STATUS, self::EVENT_TYPE_PHYSICAL_ADD, self::EVENT_TYPE_PHYSICAL_DELETE)
+					),
+					self::REPORT_FILTER_DIMENSION => self::DIMENSION_PARTNER_ID,
+					self::REPORT_GRAPH_METRICS => array(self::METRIC_ENTRIES_TOTAL),
+					self::REPORT_GRAPH_ACCUMULATE_FUNC => 'self::addAggregatedEntriesGraphs',
+				),
+			),
+			self::REPORT_GRAPH_AGGR_FUNC => 'self::aggregateUsageData'
+		),
+
 		myReportsMgr::REPORT_TYPE_ADMIN_CONSOLE => array(
 			self::REPORT_SKIP_PARTNER_FILTER => true,
 			self::REPORT_DIMENSION => self::DIMENSION_PARTNER_ID,
@@ -733,6 +853,82 @@ class kKavaReportsMgr extends kKavaBase
 			self::REPORT_GRAPH_AGGR_FUNC => 'self::aggregateUsageData',
 			self::REPORT_TOTAL_FROM_TABLE_FUNC => 'self::getTotalPeakStorageFromTable',
 		),
+			
+		// Note: historically this report returns the bandwidth in kb in table, and in mb in graph
+		myReportsMgr::REPORT_TYPE_PARTNER_BANDWIDTH_USAGE => array(
+			self::REPORT_EDIT_FILTER_FUNC => 'self::partnerUsageEditFilter',
+			self::REPORT_JOIN_REPORTS => array(
+				array(
+					self::REPORT_DATA_SOURCE => self::DATASOURCE_BANDWIDTH_USAGE,
+					self::REPORT_GRAPH_METRICS => array(self::METRIC_BANDWIDTH_SIZE_MB, self::METRIC_BANDWIDTH_SIZE_KB),
+				),
+				
+				array(
+					self::REPORT_DATA_SOURCE => self::DATASOURCE_STORAGE_USAGE,
+					self::REPORT_FILTER => array(		// can exclude logical deltas in this report
+						self::DRUID_DIMENSION => self::DIMENSION_EVENT_TYPE,
+						self::DRUID_VALUES => array(self::EVENT_TYPE_STATUS, self::EVENT_TYPE_PHYSICAL_ADD, self::EVENT_TYPE_PHYSICAL_DELETE)
+					),
+					self::REPORT_GRAPH_METRICS => array(self::METRIC_STORAGE_ADDED_MB, self::METRIC_STORAGE_DELETED_MB),
+				),
+				
+				array(
+					self::REPORT_DATA_SOURCE => self::DATASOURCE_STORAGE_USAGE,
+					self::REPORT_INTERVAL => self::INTERVAL_BASE_TO_START,
+					self::REPORT_FILTER => array(		// can exclude logical deltas in this report
+						self::DRUID_DIMENSION => self::DIMENSION_EVENT_TYPE,
+						self::DRUID_VALUES => array(self::EVENT_TYPE_STATUS, self::EVENT_TYPE_PHYSICAL_ADD, self::EVENT_TYPE_PHYSICAL_DELETE)
+					),
+					self::REPORT_GRAPH_METRICS => array(self::METRIC_STORAGE_TOTAL_MB),
+					self::REPORT_GRAPH_ACCUMULATE_FUNC => 'self::addAggregatedStorageGraphs',
+				),
+			),
+			self::REPORT_GRAPH_AGGR_FUNC => 'self::aggregateUsageData',
+			self::REPORT_GRAPH_MAP => array(
+				'bandwidth' => self::METRIC_BANDWIDTH_SIZE_MB,
+			),
+			self::REPORT_TABLE_MAP => array(
+				'avg_continuous_aggr_storage_mb' => self::METRIC_AVERAGE_STORAGE_MB,
+				'sum_partner_bandwidth_kb' => self::METRIC_BANDWIDTH_SIZE_KB,
+			),
+			self::REPORT_TABLE_FINALIZE_FUNC => 'self::addRollupRow',
+		),
+
+		myReportsMgr::REPORT_TYPE_PARTNER_USAGE_DASHBOARD => array(
+			self::REPORT_EDIT_FILTER_FUNC => 'self::partnerUsageEditFilter',
+			self::REPORT_JOIN_REPORTS => array(
+				array(
+					self::REPORT_DATA_SOURCE => self::DATASOURCE_BANDWIDTH_USAGE,
+					self::REPORT_GRAPH_METRICS => array(self::METRIC_BANDWIDTH_SIZE_KB),
+				),
+				
+				array(
+					self::REPORT_DATA_SOURCE => self::DATASOURCE_STORAGE_USAGE,
+					self::REPORT_FILTER => array(		// can exclude logical deltas in this report
+						self::DRUID_DIMENSION => self::DIMENSION_EVENT_TYPE,
+						self::DRUID_VALUES => array(self::EVENT_TYPE_STATUS, self::EVENT_TYPE_PHYSICAL_ADD, self::EVENT_TYPE_PHYSICAL_DELETE)
+					),
+					self::REPORT_GRAPH_METRICS => array(self::METRIC_STORAGE_ADDED_MB, self::METRIC_STORAGE_DELETED_MB),
+				),
+				
+				array(
+					self::REPORT_DATA_SOURCE => self::DATASOURCE_STORAGE_USAGE,
+					self::REPORT_INTERVAL => self::INTERVAL_BASE_TO_START,
+					self::REPORT_FILTER => array(		// can exclude logical deltas in this report
+						self::DRUID_DIMENSION => self::DIMENSION_EVENT_TYPE,
+						self::DRUID_VALUES => array(self::EVENT_TYPE_STATUS, self::EVENT_TYPE_PHYSICAL_ADD, self::EVENT_TYPE_PHYSICAL_DELETE)
+					),
+					self::REPORT_GRAPH_METRICS => array(self::METRIC_STORAGE_TOTAL_MB),
+					self::REPORT_GRAPH_ACCUMULATE_FUNC => 'self::addAggregatedStorageGraphs',
+				),
+			),
+			self::REPORT_GRAPH_AGGR_FUNC => 'self::aggregateUsageData',
+			self::REPORT_TABLE_MAP => array(
+				'avg_continuous_aggr_storage_mb' => self::METRIC_AVERAGE_STORAGE_MB,
+				'sum_partner_bandwidth_kb' => self::METRIC_BANDWIDTH_SIZE_KB,
+			),
+			self::REPORT_TABLE_FINALIZE_FUNC => 'self::addRollupRow',
+		),
 	);
 	
 	protected static $event_type_count_aggrs = array(
@@ -799,6 +995,7 @@ class kKavaReportsMgr extends kKavaBase
 		self::METRIC_COUNT_ADMIN => self::METRIC_COUNT_ADMIN,
 		self::METRIC_STORAGE_TOTAL_MB => self::METRIC_STORAGE_TOTAL_MB,
 		self::METRIC_BANDWIDTH_SIZE_MB => self::METRIC_BANDWIDTH_SIZE_MB,
+		self::METRIC_BANDWIDTH_SIZE_KB => self::METRIC_BANDWIDTH_SIZE_KB,
 		self::METRIC_TRANSCODING_SIZE_MB => self::METRIC_TRANSCODING_SIZE_MB,
 		self::METRIC_STORAGE_ADDED_MB => self::METRIC_STORAGE_ADDED_MB,
 		self::METRIC_STORAGE_DELETED_MB => self::METRIC_STORAGE_DELETED_MB,
@@ -1121,6 +1318,11 @@ class kKavaReportsMgr extends kKavaBase
 			self::DRUID_AGGR => array(self::METRIC_BANDWIDTH_SIZE_BYTES),
 			self::DRUID_POST_AGGR => self::getConstantRatioPostAggr(
 				self::METRIC_BANDWIDTH_SIZE_MB, self::METRIC_BANDWIDTH_SIZE_BYTES, '1048576'));
+
+		self::$metrics_def[self::METRIC_BANDWIDTH_SIZE_KB] = array(
+			self::DRUID_AGGR => array(self::METRIC_BANDWIDTH_SIZE_BYTES),
+			self::DRUID_POST_AGGR => self::getConstantRatioPostAggr(
+				self::METRIC_BANDWIDTH_SIZE_KB, self::METRIC_BANDWIDTH_SIZE_BYTES, '1024'));
 		
 		self::$metrics_def[self::METRIC_TRANSCODING_SIZE_MB] = array(
 			self::DRUID_AGGR => array(self::METRIC_FLAVOR_SIZE_BYTES),
@@ -1305,7 +1507,7 @@ class kKavaReportsMgr extends kKavaBase
 		for (;;)
 		{
 			$cur = $date->format('Ymd');
-			if (strcmp($cur, $to_day) > 0 || count($result) >= 1000)
+			if (strcmp($cur, $to_day) > 0 || count($result) >= 5000)
 			{
 				break;
 			}
@@ -1318,6 +1520,24 @@ class kKavaReportsMgr extends kKavaBase
 	}
 	
 	/// common query functions
+	protected static function shouldUseKava($partner_id, $report_type) 
+	{
+	    if ($report_type < 0 ||		// kava custom reports
+	    	in_array($report_type, self::$kava_forced_reports))
+	    {
+	    	return true;
+	    }
+	    
+	    if (in_array($report_type, self::$kava_enabled_reports) && 
+	    	kKavaBase::isPartnerAllowed($partner_id, kKavaBase::VOD_ALLOWED_PARTNERS))
+	    {
+	        return true;
+	    }
+	    
+	    KalturaLog::debug("Not using kava - partner id [$partner_id] report type [$report_type]");
+	    return false;
+	}
+		
 	protected static function toSafeId($name)
 	{
 		$name = strtoupper($name);
@@ -2353,6 +2573,11 @@ class kKavaReportsMgr extends kKavaBase
 		
 	public static function getGraph($partner_id, $report_type, reportsInputFilter $input_filter, $dimension = null, $object_ids = null)
 	{
+		if (!self::shouldUseKava($partner_id, $report_type))
+		{
+			return myReportsMgr::getGraph($partner_id, $report_type, $input_filter, $dimension, $object_ids);
+		}
+		
 		self::init();
 		
 		$report_def = self::getReportDef($report_type);
@@ -2384,6 +2609,17 @@ class kKavaReportsMgr extends kKavaBase
 			$storage[self::METRIC_AVERAGE_STORAGE_MB][$date] = $cur_storage;
 			$storage[self::METRIC_PEAK_STORAGE_MB][$date] = $cur_storage;
 			$cur_storage -= $storage[self::METRIC_STORAGE_DELETED_MB][$date];
+		}
+	}
+
+	protected static function addAggregatedEntriesGraphs(&$entries, $base_count, $dates)
+	{
+		$cur_peak = $base_count;
+		foreach ($dates as $date)
+		{
+			$cur_peak += $entries[self::METRIC_ENTRIES_ADDED][$date];
+			$entries[self::METRIC_PEAK_ENTRIES][$date] = $cur_peak;
+			$cur_peak -= $entries[self::METRIC_ENTRIES_DELETED][$date];
 		}
 	}
 	
@@ -2421,7 +2657,8 @@ class kKavaReportsMgr extends kKavaBase
 				case self::METRIC_AVERAGE_STORAGE_MB:
 					$value = $values ? array_sum($values) / count($values) : 0;
 					break;
-		
+
+				case self::METRIC_PEAK_ENTRIES:
 				case self::METRIC_PEAK_STORAGE_MB:
 					$value = $values ? max($values) : 0;
 					break;
@@ -2982,8 +3219,7 @@ class kKavaReportsMgr extends kKavaBase
 		return $total_count;
 	}
 
-	// TODO: remove the report_type param once legacy reports are removed
-	protected static function getSimpleTableImpl($partner_id, $report_type, $report_def, 
+	protected static function getSimpleTableImpl($partner_id, $report_def, 
 		reportsInputFilter $input_filter,
 		$page_size, $page_index, $order_by, $object_ids = null, $flags = 0)
 	{
@@ -2995,23 +3231,6 @@ class kKavaReportsMgr extends kKavaBase
 		$druid_filter = self::getDruidFilter($partner_id, $report_def, $input_filter, $object_ids);
 		$dimension = self::getDimension($report_def, $object_ids);
 		$metrics = self::getMetrics($report_def);
-
-		if (array_key_exists(self::REPORT_LEGACY, $report_def))
-		{
-			//get the report from the legacy dwh
-			list($headers, $data, $total_count) = myReportsMgr::getTable($partner_id, $report_type, $input_filter, $page_size, $page_index, $order_by, $object_ids, null);
-
-			switch ($report_type)
-			{
-				case myReportsMgr::REPORT_TYPE_VPAAS_USAGE:
-					if ($object_ids)
-						throw new Exception('objectIds filter is not supported for this report');
-
-					$data = self::getVpaasUsageReport($report_def, $data, $headers, $partner_id, $intervals, $metrics, $druid_filter, $input_filter->timeZoneOffset);
-					break;
-			}
-			return array($headers, $data, $total_count);
-		}
 
 		if (!$metrics)
 		{
@@ -3354,7 +3573,7 @@ class kKavaReportsMgr extends kKavaBase
 		return array($headers, $data, $total_count);
 	}
 
-	protected static function getJoinTableImpl($partner_id, $report_type, $report_def,
+	protected static function getJoinTableImpl($partner_id, $report_def,
 			reportsInputFilter $input_filter,
 			$page_size, $page_index, $order_by, $object_ids = null, $flags = 0)
 	{
@@ -3390,7 +3609,7 @@ class kKavaReportsMgr extends kKavaBase
 		$cur_report_def[self::REPORT_DIMENSION] = $report_def[self::REPORT_DIMENSION];
 		$cur_report_def[self::REPORT_DIMENSION_HEADERS] = array('dimension');
 		
-		$result = self::getTableImpl($partner_id, $report_type, $cur_report_def, $input_filter,
+		$result = self::getTableImpl($partner_id, $cur_report_def, $input_filter,
 			$page_size * $page_index, 1, $order_by, $object_ids, $flags);
 		
 		$headers = array_merge(
@@ -3426,7 +3645,7 @@ class kKavaReportsMgr extends kKavaBase
 			if (count($rows) < $page_size)
 			{				
 				// using a single topN - the metrics of this query are not used
-				$result = self::getTableImpl($partner_id, $report_type, $cur_report_def, $input_filter,
+				$result = self::getTableImpl($partner_id, $cur_report_def, $input_filter,
 					$page_size * $page_index, 1, $order_by, $object_ids, $flags | self::GET_TABLE_FLAG_IDS_ONLY);
 								
 				foreach ($result[1] as $row)
@@ -3480,7 +3699,7 @@ class kKavaReportsMgr extends kKavaBase
 				$cur_report_def[self::REPORT_FILTER_DIMENSION] = $report_def[self::REPORT_DIMENSION];;
 				$cur_report_def[self::REPORT_DIMENSION_HEADERS] = array('dimension');
 				
-				$result = self::getTableImpl($partner_id, $report_type, $cur_report_def, $input_filter,
+				$result = self::getTableImpl($partner_id, $cur_report_def, $input_filter,
 					count($ids_to_get), 1, null, implode(',', array_keys($ids_to_get)), $flags);
 				
 				foreach ($result[1] as $row)
@@ -3511,12 +3730,17 @@ class kKavaReportsMgr extends kKavaBase
 		}
 		
 		return array($headers, $data, $total_count);
-	}		
+	}
 
-	protected static function getTableImpl($partner_id, $report_type, $report_def, 
+	protected static function getTableImpl($partner_id, $report_def, 
 		reportsInputFilter $input_filter,
 		$page_size, $page_index, $order_by, $object_ids = null, $flags = 0)
 	{
+		if (isset($report_def[self::REPORT_EDIT_FILTER_FUNC]))
+		{
+			call_user_func($report_def[self::REPORT_EDIT_FILTER_FUNC], $input_filter);
+		}
+				
 		if (!isset($report_def[self::REPORT_DIMENSION]))
 		{
 			$result = self::getGraphImpl($partner_id, $report_def, $input_filter, $object_ids);
@@ -3529,12 +3753,12 @@ class kKavaReportsMgr extends kKavaBase
 		}
 		else if (isset($report_def[self::REPORT_JOIN_REPORTS]))
 		{
-			$result = self::getJoinTableImpl($partner_id, $report_type, $report_def, $input_filter, 
+			$result = self::getJoinTableImpl($partner_id, $report_def, $input_filter, 
 				$page_size, $page_index, $order_by, $object_ids, $flags);
 		}
 		else 
 		{
-			$result = self::getSimpleTableImpl($partner_id, $report_type, $report_def, $input_filter,
+			$result = self::getSimpleTableImpl($partner_id, $report_def, $input_filter,
 				$page_size, $page_index, $order_by, $object_ids, $flags);
 		}
 		
@@ -3608,8 +3832,14 @@ class kKavaReportsMgr extends kKavaBase
 	}
 	
 	public static function getTable($partner_id, $report_type, reportsInputFilter $input_filter,
-		$page_size, $page_index, $order_by, $object_ids = null, $isCsv = false)
+		$page_size, $page_index, $order_by, $object_ids = null, $offset = null, $isCsv = false)
 	{
+		if (!self::shouldUseKava($partner_id, $report_type))
+		{
+			return myReportsMgr::getTable($partner_id, $report_type, $input_filter,
+				$page_size, $page_index, $order_by, $object_ids, $offset);
+		}
+		
 		self::init();		
 
 		// pager
@@ -3629,7 +3859,7 @@ class kKavaReportsMgr extends kKavaBase
 		
 		$flags = $isCsv ? self::GET_TABLE_FLAG_IS_CSV : 0;
 		
-		$result = self::getTableImpl($partner_id, $report_type, $report_def, $input_filter,
+		$result = self::getTableImpl($partner_id, $report_def, $input_filter,
 			$page_size, $page_index, $order_by, $object_ids, $flags);
 	
 		// reorder
@@ -3652,6 +3882,23 @@ class kKavaReportsMgr extends kKavaBase
 		return $result;
 	}
 	
+	protected static function partnerUsageEditFilter($input_filter)
+	{
+		$current_date_id = date('Ymd');
+		$from_day = min($input_filter->from_day, $current_date_id);
+		
+		$month_start = substr($from_day, 0, 6) . '01';
+		$date = self::dateIdToDateTime($month_start);
+		$date->modify('+1 month');
+		$date->modify('-1 day');
+		$month_end = min($date->format('Ymd'), $current_date_id);
+		
+		$is_free_package = $input_filter->extra_map[myPartnerUtils::IS_FREE_PACKAGE_PLACE_HOLDER] == 'TRUE';
+		$input_filter->from_day = $is_free_package ? str_replace('-', '', self::BASE_DATE_ID) : $month_start;
+		$input_filter->to_day = $month_end;
+		$input_filter->interval = reportInterval::MONTHS;
+	}
+	
 	protected static function addCombinedUsageColumn(&$result)
 	{
 		$headers = $result[0];
@@ -3668,7 +3915,36 @@ class kKavaReportsMgr extends kKavaBase
 			$row[] = $row[$bandwidth] + $row[$storage]; 
 		}
 	}
+
+	protected static function getRollupRow($data)
+	{
+		$row = reset($data);
+		for (;;)
+		{
+			$cur_row = next($data);
+			if (!$cur_row)
+			{
+				break;
+			}
+			
+			for ($i = 1; $i < count($row); $i++)		// starting from 1 to skip the date id
+			{
+				$row[$i] += $cur_row[$i];
+			}
+		}
 		
+		return $row;
+	}
+	
+	protected static function addRollupRow(&$result)
+	{
+		list($headers, $data, $total_count) = $result;
+		
+		$data[] = self::getRollupRow($data);
+		
+		$result = array($headers, $data, $total_count);
+	}
+
 	/// total functions
 	protected static function getTotalPeakStorageFromTable($table)
 	{
@@ -3771,7 +4047,7 @@ class kKavaReportsMgr extends kKavaBase
 		if (isset($report_def[self::REPORT_TOTAL_FROM_TABLE_FUNC]))
 		{
 			$input_filter->interval = self::INTERVAL_ALL;
-			$table = self::getTableImpl($partner_id, 0, $report_def, $input_filter, self::MAX_RESULT_SIZE, 1, null, $object_ids);
+			$table = self::getTableImpl($partner_id, $report_def, $input_filter, self::MAX_RESULT_SIZE, 1, null, $object_ids);
 			$result = call_user_func($report_def[self::REPORT_TOTAL_FROM_TABLE_FUNC], $table);
 		}
 		else if (!isset($report_def[self::REPORT_DIMENSION]) || isset($report_def[self::REPORT_JOIN_GRAPHS]))
@@ -3794,6 +4070,11 @@ class kKavaReportsMgr extends kKavaBase
 	
 	public static function getTotal($partner_id, $report_type, reportsInputFilter $input_filter, $object_ids = null)
 	{
+		if (!self::shouldUseKava($partner_id, $report_type))
+		{
+			return myReportsMgr::getTotal($partner_id, $report_type, $input_filter, $object_ids);
+		}
+		
 		self::init();
 
 		$report_def = self::getReportDef($report_type);
@@ -3817,6 +4098,11 @@ class kKavaReportsMgr extends kKavaBase
 
 	public static function getBaseTotal($partner_id, $report_type, reportsInputFilter $input_filter, $object_ids = null )
 	{
+		if (!self::shouldUseKava($partner_id, $report_type))
+		{
+			return myReportsMgr::getBaseTotal($partner_id, $report_type, $input_filter, $object_ids);
+		}
+		
 		switch ($report_type)
 		{
 			case myReportsMgr::REPORT_TYPE_USER_USAGE:
@@ -3904,6 +4190,7 @@ class kKavaReportsMgr extends kKavaBase
 				1,
 				$report_def['order_by'],
 				$object_ids,
+				null,
 				true);
 		}
 		else
@@ -3924,34 +4211,6 @@ class kKavaReportsMgr extends kKavaBase
 		return array($header, $data);
 	}
 
-	protected static function getVpaasUsageReport($report_def, $data, $headers, $partner_id, $intervals, $metrics, $druid_filter, $timezone_offset)
-	{
-		$granularity = $report_def[self::REPORT_GRANULARITY];
-		$granularity_def = self::getGranularityDef($granularity, $timezone_offset);
-		$query = self::getTimeSeriesReport(null, $partner_id, $intervals, $granularity_def, $metrics, $druid_filter);
-		$result = self::runQuery($query);
-		$report_metrics_to_headers = array();
-		foreach ($metrics as $column)
-		{
-			$report_metrics_to_headers[$column] = self::$metrics_to_headers[$column];
-		}
-		$druid_result = self::getGraphsByDateId($result, $report_metrics_to_headers, $timezone_offset, self::$transform_time_dimensions[$granularity]);
-		$mapping = $report_def[self::REPORT_LEGACY][self::REPORT_LEGACY_MAPPING];
-		$druid_field = reset($mapping);
-		$legacy_field = key($mapping);
-		$legacy_field_index = array_search($legacy_field, $headers);
-
-		$key_index = array_search($report_def[self::REPORT_LEGACY][self::REPORT_LEGACY_JOIN_FIELD], $headers);
-		if (false !== $legacy_field_index)
-		{
-			foreach ($data as &$row)
-			{
-				$row[$legacy_field_index] = isset($druid_result[$druid_field][$row[$key_index]]) ? $druid_result[$druid_field][$row[$key_index]] : 0;
-			}
-		}
-		return $data;
-	}
-	
 	/// csv functions
 	protected static function getCsvData(
 		$partner_id,
@@ -3994,8 +4253,12 @@ class kKavaReportsMgr extends kKavaBase
 				$partner_id,
 				$report_type,
 				$input_filter,
-				$page_size, $page_index,
-				$order_by, $object_ids, true);
+				$page_size, 
+				$page_index,
+				$order_by, 
+				$object_ids, 
+				null, 
+				true);
 		}
 
 		$csv = myCsvReport::createReport($report_title, $report_text, $headers,
@@ -4030,6 +4293,18 @@ class kKavaReportsMgr extends kKavaBase
 		$object_ids = null,
 		$page_size =10, $page_index =0, $order_by)
 	{
+		if (!self::shouldUseKava($partner_id, $report_type))
+		{
+			return myReportsMgr::getUrlForReportAsCsv(
+				$partner_id, 
+				$report_title, $report_text, $headers, 
+				$report_type, 
+				$input_filter, 
+				$dimension, 
+				$object_ids,
+				$page_size, $page_index, $order_by);					
+		}
+		
 		list($file_path, $file_name) = myReportsMgr::createFileName($partner_id, $report_type, $input_filter, $dimension, $object_ids, $page_size, $page_index, $order_by);
 
 		$data = self::getCsvData($partner_id,