+++ api_v3/services/ReportService.php
@@ -7,25 +7,6 @@
  */
 class ReportService extends KalturaBaseService
 {
-	static $kavaReports = array(KalturaReportType::TOP_CONTENT,
-		KalturaReportType::CONTENT_DROPOFF,
-		KalturaReportType::CONTENT_INTERACTIONS,
-		KalturaReportType::MAP_OVERLAY,
-		KalturaReportType::TOP_SYNDICATION,
-		KalturaReportType::USER_ENGAGEMENT,
-		KalturaReportType::SPECIFIC_USER_ENGAGEMENT,
-		KalturaReportType::USER_TOP_CONTENT,
-		KalturaReportType::USER_CONTENT_DROPOFF,
-		KalturaReportType::USER_CONTENT_INTERACTIONS,
-		KalturaReportType::APPLICATIONS,
-		KalturaReportType::PLATFORMS,
-		KalturaReportType::OPERATING_SYSTEM,
-		KalturaReportType::BROWSERS,
-		KalturaReportType::LIVE,
-		KalturaReportType::TOP_PLAYBACK_CONTEXT,
-		KalturaReportType::VPAAS_USAGE,
-	);
-
 	public function initService($serviceId, $serviceName, $actionName)
 	{
 		parent::initService($serviceId, $serviceName, $actionName);
@@ -89,9 +70,8 @@ class ReportService extends KalturaBaseService
 		if($reportType == KalturaReportType::PARTNER_USAGE || $reportType == KalturaReportType::VAR_USAGE)
 			$objectIds = $this->validateObjectsAreAllowedPartners($objectIds);
 	
-	    $reportsMgrClass = $this->getReportsManagerClass($reportType);
-		        
-		$reportGraphs =  KalturaReportGraphArray::fromReportDataArray (call_user_func(array($reportsMgrClass, "getGraph"),$this->getPartnerId(),
+		$reportGraphs =  KalturaReportGraphArray::fromReportDataArray(kKavaReportsMgr::getGraph(
+		    $this->getPartnerId(),
 		    $reportType,
 		    $reportInputFilter->toReportsInputFilter(),
 		    $dimension,
@@ -116,9 +96,8 @@ class ReportService extends KalturaBaseService
 		
 		$reportTotal = new KalturaReportTotal();
 		
-		$reportsMgrClass = $this->getReportsManagerClass($reportType);
-		
-		list ( $header , $data ) = call_user_func(array($reportsMgrClass, "getTotal"), $this->getPartnerId() ,
+		list ( $header , $data ) = kKavaReportsMgr::getTotal(
+		    $this->getPartnerId() ,
 		    $reportType ,
 		    $reportInputFilter->toReportsInputFilter() , $objectIds);
 		
@@ -138,10 +117,8 @@ class ReportService extends KalturaBaseService
 	 */
 	public function getBaseTotalAction( $reportType , KalturaReportInputFilter $reportInputFilter , $objectIds = null )
 	{
-		$reportsMgrClass = $this->getReportsManagerClass($reportType);
-		
 		$reportSubTotals =  KalturaReportBaseTotalArray::fromReportDataArray(  
-			call_user_func(array($reportsMgrClass, "getBaseTotal"), 
+			kKavaReportsMgr::getBaseTotal( 
 				$this->getPartnerId() , 
 				$reportType , 
 				$reportInputFilter->toReportsInputFilter() ,
@@ -169,9 +146,8 @@ class ReportService extends KalturaBaseService
 		
 		$reportTable = new KalturaReportTable();
 		
-		$reportsMgrClass = $this->getReportsManagerClass($reportType);
-		
-		list ( $header , $data , $totalCount ) = call_user_func(array($reportsMgrClass, "getTable"), $this->getPartnerId() ,
+		list ( $header , $data , $totalCount ) = kKavaReportsMgr::getTable(
+		    $this->getPartnerId() ,
 		    $reportType ,
 		    $reportInputFilter->toReportsInputFilter() ,
 		    $pager->pageSize , $pager->pageIndex ,
@@ -210,9 +186,8 @@ class ReportService extends KalturaBaseService
 			$objectIds = $this->validateObjectsAreAllowedPartners($objectIds);
 		
 		try {
-			$reportsMgrClass = $this->getReportsManagerClass($reportType);
-
-			$report = call_user_func(array($reportsMgrClass, "getUrlForReportAsCsv"), $this->getPartnerId(),
+			$report = kKavaReportsMgr::getUrlForReportAsCsv(
+				$this->getPartnerId(),
 				$reportTitle,
 				$reportText,
 				$headers,
@@ -383,16 +358,5 @@ class ReportService extends KalturaBaseService
 		$partnerIdParam->key = 'partner_id';
 		$partnerIdParam->value = $this->getPartnerId();
 		$params[] = $partnerIdParam;
-	}
-	
-	protected function getReportsManagerClass($reportType) 
-	{
-	    $reportsMgrClass = "myReportsMgr";
-	    if (in_array($reportType, self::$kavaReports) && kKavaBase::isPartnerAllowed($this->getPartnerId(), kKavaBase::VOD_ALLOWED_PARTNERS))
-	    {
-	        $reportsMgrClass = "kKavaReportsMgr";
-	    }
-	    
-	    return $reportsMgrClass;
 	}
 }