+++ plugins/conference/lib/api/enums/KalturaConferenceRoomStatus.php
@@ -1,0 +1,10 @@
+<?php
+/**
+ * @package plugins.conference
+ * @subpackage api.enum
+ * @see ConferenceRoomStatus
+ */
+class KalturaConferenceRoomStatus extends KalturaEnum implements ConferenceRoomStatus
+{
+
+}