+++ plugins/conference/lib/api/filters/orderEnums/KalturaConferenceServerNodeOrderBy.php
@@ -1,0 +1,8 @@
+<?php
+/**
+ * @package plugins.conference
+ * @subpackage api.filters.enum
+ */
+class KalturaConferenceServerNodeOrderBy extends KalturaServerNodeOrderBy
+{
+}