+++ plugins/konference/lib/api/filters/orderEnums/KalturaConferenceServerNodeOrderBy.php
@@ -1,8 +1,0 @@
-<?php
-/**
- * @package plugins.konference
- * @subpackage api.filters.enum
- */
-class KalturaConferenceServerNodeOrderBy extends KalturaServerNodeOrderBy
-{
-}