+++ plugins/search/providers/elastic_search/lib/model/enum/ESearchCategoryUserFieldName.php
@@ -1,0 +1,9 @@
+<?php
+/**
+ * @package plugins.elasticSearch
+ * @subpackage api.enum
+ */
+class ESearchCategoryUserFieldName extends KalturaStringEnum
+{
+	const USER_ID = 'kuser_ids';
+}